Move InsetSquare struct into a new file and rename to ZoomedSquare
